- separarted logic from UI - save window state - exit app on error
