+++ AnnoDesigner.Core/Models/ICloseable.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AnnoDesigner.Core.Models
+{
+    public interface ICloseable
+    {
+        void Close();
+    }
+}
