+++ AnnoDesigner.Core/Models/Notify.cs
@@ -5,7 +5,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Runtime.CompilerServices;
 
-namespace AnnoDesigner.model
+namespace AnnoDesigner.Core.Models
 {
 
     /// <summary>
