+++ AnnoDesigner/InputWindow.xaml
@@ -5,6 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:l="clr-namespace:AnnoDesigner.Localization"
         xmlns:local="clr-namespace:AnnoDesigner"
+        xmlns:viewModel="clr-namespace:AnnoDesigner.viewmodel"
         mc:Ignorable="d"
         Title="InputWindow"
         WindowStartupLocation="CenterScreen"
@@ -15,7 +16,7 @@
         WindowStyle="SingleBorderWindow"
         ResizeMode="CanMinimize">
     <Window.Resources>
-        <l:MainWindow x:Key="localization"></l:MainWindow>
+        <viewModel:MainViewModel x:Key="localization"></viewModel:MainViewModel>
     </Window.Resources>
     <Window.DataContext>
         <Binding Source="{StaticResource localization}"></Binding>
