+++ AnnoDesigner/Localization/Localization.cs
@@ -1,11 +1,5 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using AnnoDesigner.model;
-using AnnoDesigner.viewmodel;
 
 namespace AnnoDesigner.Localization
 {
@@ -602,916 +596,6 @@ namespace AnnoDesigner.Localization
         {
 
         }
-
-
-    }
-
-    //These classes cannot be nested as they are used for data binding
-
-    public class About : Notify
-    {
-        public About()
-        {
-            UpdateLanguage();
-        }
-
-        public void UpdateLanguage()
-        {
-            string language = Localization.GetLanguageCodeFromName(AnnoDesigner.MainWindow.SelectedLanguage);
-
-            TitleAbout = Localization.Translations[language]["TitleAbout"];
-            Title = Localization.Translations[language]["Title"];
-            //Title = "Modified", ModifiedAnnoDesigner = "Modified" + " Anno Designer"
-            ModifiedAnnoDesigner = Localization.Translations[language]["Title"];
-            BuildingLayoutDesigner = Localization.Translations[language]["BuildingLayoutDesigner"];
-
-            //Credits
-            Credits = Localization.Translations[language]["Credits"];
-            OriginalApplicationBy = Localization.Translations[language]["OriginalApplicationBy"];
-            BuildingPresets = Localization.Translations[language]["BuildingPresets"];
-            CombinedForAnnoVersions = Localization.Translations[language]["CombinedForAnnoVersions"];
-            AdditionalChanges = Localization.Translations[language]["AdditionalChanges"];
-            ManyThanks = Localization.Translations[language]["ManyThanks"];
-            VisitTheFandom = Localization.Translations[language]["VisitTheFandom"];
-            OriginalHomepage = Localization.Translations[language]["OriginalHomepage"];
-            ProjectHomepage = Localization.Translations[language]["ProjectHomepage"];
-            GoToFandom = Localization.Translations[language]["GoToFandom"];
-            Close = Localization.Translations[language]["Close"];
-        }
-
-        //Generated from:
-        //...
-        //public string Prop1 {get; set;}
-        //public string Prop2 {get; set;}
-        //...
-        //find expr: public (string) (.+?) {.+
-        //With the following regex (in a compatible editor that supports lowercasing of values
-        //within regex expressions):
-        //private $1 _\l$2; \r\n public $1 $2 \r\n { \r\n get { return _\l$2; } \r\n set \r\n { \r\n UpdateProperty\(ref _\l$2, value\); \r\n}\r\n}
-
-        private string _titleAbout;
-        public string TitleAbout
-        {
-            get { return _titleAbout; }
-            set
-            {
-                UpdateProperty(ref _titleAbout, value);
-            }
-        }
-        private string _title;
-        public string Title
-        {
-            get { return TitleAbout + " Anno Designer (" + _title + ")"; }
-            set
-            {
-                UpdateProperty(ref _title, value);
-            }
-        }
-
-        //Setter not needed
-        private string _modifiedAnnoDesigner;
-        public string ModifiedAnnoDesigner
-        {
-            get { return _modifiedAnnoDesigner + " Anno Designer"; }
-            set
-            {
-                UpdateProperty(ref _modifiedAnnoDesigner, value);
-            }
-        }
-        private string _buildingLayoutDesigner;
-        public string BuildingLayoutDesigner
-        {
-            get { return _buildingLayoutDesigner; }
-            set
-            {
-                UpdateProperty(ref _buildingLayoutDesigner, value);
-            }
-        }
-
-        //Credits
-        private string _credits;
-        public string Credits
-        {
-            get { return _credits; }
-            set
-            {
-                UpdateProperty(ref _credits, value);
-            }
-        }
-        private string _originalApplicationBy;
-        public string OriginalApplicationBy
-        {
-            get { return _originalApplicationBy; }
-            set
-            {
-                UpdateProperty(ref _originalApplicationBy, value);
-            }
-        }
-        private string _buildingPresets;
-        public string BuildingPresets
-        {
-            get { return _buildingPresets; }
-            set
-            {
-                UpdateProperty(ref _buildingPresets, value);
-            }
-        }
-        private string _combinedForAnnoVersions;
-        public string CombinedForAnnoVersions
-        {
-            get { return _combinedForAnnoVersions; }
-            set
-            {
-                UpdateProperty(ref _combinedForAnnoVersions, value);
-            }
-        }
-        private string _additionalChanges;
-        public string AdditionalChanges
-        {
-            get { return _additionalChanges; }
-            set
-            {
-                UpdateProperty(ref _additionalChanges, value);
-            }
-        }
-        private string _manyThanks;
-        public string ManyThanks
-        {
-            get { return _manyThanks; }
-            set
-            {
-                UpdateProperty(ref _manyThanks, value);
-            }
-        }
-        private string _visitTheFandom;
-        public string VisitTheFandom
-        {
-            get { return _visitTheFandom; }
-            set
-            {
-                UpdateProperty(ref _visitTheFandom, value);
-            }
-        }
-        private string _originalHomepage;
-        public string OriginalHomepage
-        {
-            get { return _originalHomepage; }
-            set
-            {
-                UpdateProperty(ref _originalHomepage, value);
-            }
-        }
-        private string _projectHomepage;
-        public string ProjectHomepage
-        {
-            get { return _projectHomepage; }
-            set
-            {
-                UpdateProperty(ref _projectHomepage, value);
-            }
-        }
-        private string _goToFandom;
-        public string GoToFandom
-        {
-            get { return _goToFandom; }
-            set
-            {
-                UpdateProperty(ref _goToFandom, value);
-            }
-        }
-        private string _close;
-        public string Close
-        {
-            get { return _close; }
-            set
-            {
-                UpdateProperty(ref _close, value);
-            }
-        }
-
-
-    }
-
-    public class AnnoCanvas : Notify
-    {
-        public AnnoCanvas()
-        {
-            UpdateLanguage();
-        }
-
-        public void UpdateLanguage()
-        {
-            string language = Localization.GetLanguageCodeFromName(AnnoDesigner.MainWindow.SelectedLanguage);
-
-            //Statistics Section
-            NothingPlaced = Localization.Translations[language]["StatNothingPlaced"];
-            BoundingBox = Localization.Translations[language]["StatBoundingBox"];
-            MinimumArea = Localization.Translations[language]["StatMinimumArea"];
-            SpaceEfficiency = Localization.Translations[language]["StatSpaceEfficiency"];
-
-        }
-        //Statistics Section
-        private string _nothingPlaced;
-        public string NothingPlaced
-        {
-            get { return _nothingPlaced; }
-            set
-            {
-                UpdateProperty(ref _nothingPlaced, value);
-            }
-        }
-        private string _boundingBox;
-        public string BoundingBox
-        {
-            get { return _boundingBox; }
-            set
-            {
-                UpdateProperty(ref _boundingBox, value);
-            }
-        }
-        private string _minimumArea;
-        public string MinimumArea
-        {
-            get { return _minimumArea; }
-            set
-            {
-                UpdateProperty(ref _minimumArea, value);
-            }
-        }
-        private string _spaceEfficiency;
-        public string SpaceEfficiency
-        {
-            get { return _spaceEfficiency; }
-            set
-            {
-                UpdateProperty(ref _spaceEfficiency, value);
-            }
-        }
-    }
-
-    ////Probably nothing to add in here
-    //public class App : Notify
-    //{
-    //}
-
-
-    /// <summary>
-    /// Holds information about the current localized symbols for the "MainWindow.xaml" window
-    /// </summary>
-    public class MainWindow : Notify
-    {
-        public MainWindow()
-        {
-            _statisticsViewModel = new StatisticsViewModel();
-            _buildingSettingsViewModel = new BuildingSettingsViewModel();
-            _presetTreeViewModel = new PresetsTreeViewModel(new TreeLocalization());
-
-            UpdateLanguage();
-        }
-
-        public void UpdateLanguage()
-        {
-            string language = Localization.GetLanguageCodeFromName(AnnoDesigner.MainWindow.SelectedLanguage);
-
-            StatisticsViewModel.TextNothingPlaced = Localization.Translations[language]["StatNothingPlaced"];
-            StatisticsViewModel.TextBoundingBox = Localization.Translations[language]["StatBoundingBox"];
-            StatisticsViewModel.TextMinimumArea = Localization.Translations[language]["StatMinimumArea"];
-            StatisticsViewModel.TextSpaceEfficiency = Localization.Translations[language]["StatSpaceEfficiency"];
-            StatisticsViewModel.TextBuildings = Localization.Translations[language]["StatBuildings"];
-            StatisticsViewModel.TextBuildingsSelected = Localization.Translations[language]["StatBuildingsSelected"];
-            StatisticsViewModel.TextTiles = Localization.Translations[language]["StatTiles"];
-            StatisticsViewModel.TextNameNotFound = Localization.Translations[language]["StatNameNotFound"];
-
-            Ok = Localization.Translations[language]["OK"];
-            Cancel = Localization.Translations[language]["Cancel"];
-
-            //File Menu
-            File = Localization.Translations[language]["File"];
-            NewCanvas = Localization.Translations[language]["NewCanvas"];
-            Open = Localization.Translations[language]["Open"];
-            Save = Localization.Translations[language]["Save"];
-            SaveAs = Localization.Translations[language]["SaveAs"];
-            CopyLayoutToClipboard = Localization.Translations[language]["CopyLayoutToClipboard"];
-            LoadLayoutFromJson = Localization.Translations[language]["LoadLayoutFromJson"];
-            Exit = Localization.Translations[language]["Exit"];
-
-            //Extras Menu
-            Extras = Localization.Translations[language]["Extras"];
-            Normalize = Localization.Translations[language]["Normalize"];
-            ResetZoom = Localization.Translations[language]["ResetZoom"];
-            RegisterFileExtension = Localization.Translations[language]["RegisterFileExtension"];
-            UnregisterFileExtension = Localization.Translations[language]["UnregisterFileExtension"];
-
-            //Export Menu
-            Export = Localization.Translations[language]["Export"];
-            ExportImage = Localization.Translations[language]["ExportImage"];
-            UseCurrentZoomOnExportedImage = Localization.Translations[language]["UseCurrentZoomOnExportedImage"];
-            RenderSelectionHighlightsOnExportedImage = Localization.Translations[language]["RenderSelectionHighlightsOnExportedImage"];
-
-            //Manage Stats Menu
-            ManageStats = Localization.Translations[language]["ManageStats"];
-            ShowStats = Localization.Translations[language]["ShowStats"];
-            BuildingCount = Localization.Translations[language]["BuildingCount"];
-
-            //Language Menu
-            Language = Localization.Translations[language]["Language"];
-
-            //Help Menu
-            Help = Localization.Translations[language]["Help"];
-            Version = Localization.Translations[language]["Version"];
-            FileVersion = Localization.Translations[language]["FileVersion"];
-            PresetsVersion = Localization.Translations[language]["PresetsVersion"];
-            CheckForUpdates = Localization.Translations[language]["CheckForUpdates"];
-            EnableAutomaticUpdateCheck = Localization.Translations[language]["EnableAutomaticUpdateCheck"];
-            GoToProjectHomepage = Localization.Translations[language]["GoToProjectHomepage"];
-            OpenWelcomePage = Localization.Translations[language]["OpenWelcomePage"];
-            AboutAnnoDesigner = Localization.Translations[language]["AboutAnnoDesigner"];
-
-            //Other
-            ShowGrid = Localization.Translations[language]["ShowGrid"];
-            ShowLabels = Localization.Translations[language]["ShowLabels"];
-            ShowIcons = Localization.Translations[language]["ShowIcons"];
-
-            //DockPanel
-            BuildingSettingsViewModel.TextHeader = Localization.Translations[language]["BuildingSettings"];
-            BuildingSettingsViewModel.TextSize = Localization.Translations[language]["Size"];
-            BuildingSettingsViewModel.TextColor = Localization.Translations[language]["Color"];
-            BuildingSettingsViewModel.TextBuildingName = Localization.Translations[language]["Label"];
-            BuildingSettingsViewModel.TextIcon = Localization.Translations[language]["Icon"];
-            BuildingSettingsViewModel.TextInfluenceType = Localization.Translations[language]["InfluenceType"];
-            None = Localization.Translations[language]["None"];
-            BuildingSettingsViewModel.TextRadius = Localization.Translations[language]["Radius"];
-            BuildingSettingsViewModel.TextDistance = Localization.Translations[language]["Distance"];
-            Both = Localization.Translations[language]["Both"];
-            BuildingSettingsViewModel.TextPavedStreet = Localization.Translations[language]["PavedStreet"];
-            BuildingSettingsViewModel.TextPavedStreetWarningTitle = Localization.Translations[language]["PavedStreetWarningTitle"];
-            BuildingSettingsViewModel.TextPavedStreetToolTip = Localization.Translations[language]["PavedStreetToolTip"];
-            BuildingSettingsViewModel.TextOptions = Localization.Translations[language]["Options"];
-            BuildingSettingsViewModel.TextEnableLabel = Localization.Translations[language]["EnableLabel"];
-            BuildingSettingsViewModel.TextBorderless = Localization.Translations[language]["Borderless"];
-            BuildingSettingsViewModel.TextRoad = Localization.Translations[language]["Road"];
-            BuildingSettingsViewModel.TextPlaceBuilding = Localization.Translations[language]["PlaceBuilding"];
-            Search = Localization.Translations[language]["Search"];
-            TreeViewSearchToolTip = Localization.Translations[language]["SearchToolTip"];
-            BuildingSettingsViewModel.TextApplyColorToSelection = Localization.Translations[language]["ApplyColorToSelection"];
-            BuildingSettingsViewModel.TextApplyColorToSelectionToolTip = Localization.Translations[language]["ApplyColorToSelectionToolTip"];
-            BuildingSettingsViewModel.TextApplyPredefinedColorToSelection = Localization.Translations[language]["ApplyPredefinedColorToSelection"];
-            BuildingSettingsViewModel.TextApplyPredefinedColorToSelectionToolTip = Localization.Translations[language]["ApplyPredefinedColorToSelectionToolTip"];
-            BuildingSettingsViewModel.TextAvailableColors = Localization.Translations[language]["AvailableColors"];
-            BuildingSettingsViewModel.TextStandardColors = Localization.Translations[language]["StandardColors"];
-            BuildingSettingsViewModel.TextRecentColors = Localization.Translations[language]["RecentColors"];
-            BuildingSettingsViewModel.TextStandard = Localization.Translations[language]["Standard"];
-            BuildingSettingsViewModel.TextAdvanced = Localization.Translations[language]["Advanced"];
-            BuildingSettingsViewModel.TextColorsInLayout = Localization.Translations[language]["ColorsInLayout"];
-            BuildingSettingsViewModel.TextColorsInLayoutToolTip = Localization.Translations[language]["ColorsInLayoutToolTip"];
-
-            //Status Bar
-            StatusBarControls = Localization.Translations[language]["StatusBarControls"];
-            StatusBarItemsOnClipboard = Localization.Translations[language]["StatusBarItemsOnClipboard"];
-        }
-
-        //Generated from:
-        //...
-        //public string Prop1 {get; set;}
-        //public string Prop2 {get; set;}
-        //...
-        //find expr: public (string) (.+?) {.+
-        //With the following regex (in a compatible editor that supports lowercasing of values
-        //within regex expressions):
-        //private $1 _\l$2; \r\n public $1 $2 \r\n { \r\n get { return _\l$2; } \r\n set \r\n { \r\n UpdateProperty\(ref _\l$2, value\); \r\n}\r\n}
-
-        #region File Menu
-
-        private string _file;
-        public string File
-        {
-            get { return _file; }
-            set
-            {
-                UpdateProperty(ref _file, value);
-            }
-        }
-
-        private string _newCanvas;
-        public string NewCanvas
-        {
-            get { return _newCanvas; }
-            set
-            {
-                UpdateProperty(ref _newCanvas, value);
-            }
-        }
-
-        private string _open;
-        public string Open
-        {
-            get { return _open; }
-            set
-            {
-                UpdateProperty(ref _open, value);
-            }
-        }
-
-        private string _save;
-        public string Save
-        {
-            get { return _save; }
-            set
-            {
-                UpdateProperty(ref _save, value);
-            }
-        }
-
-        private string _saveAs;
-        public string SaveAs
-        {
-            get { return _saveAs; }
-            set
-            {
-                UpdateProperty(ref _saveAs, value);
-            }
-        }
-
-        private string _copyLayoutToClipboard;
-        public string CopyLayoutToClipboard
-        {
-            get { return _copyLayoutToClipboard; }
-            set
-            {
-                UpdateProperty(ref _copyLayoutToClipboard, value);
-            }
-        }
-
-        private string _loadLayoutFromJson;
-        public string LoadLayoutFromJson
-        {
-            get { return _loadLayoutFromJson; }
-            set
-            {
-                UpdateProperty(ref _loadLayoutFromJson, value);
-            }
-        }
-
-        private string _ok;
-        public string Ok
-        {
-            get { return _ok; }
-            set
-            {
-                UpdateProperty(ref _ok, value);
-            }
-        }
-
-        private string _cancel;
-        public string Cancel
-        {
-            get { return _cancel; }
-            set
-            {
-                UpdateProperty(ref _cancel, value);
-            }
-        }
-
-        private string _exit;
-        public string Exit
-        {
-            get { return _exit; }
-            set
-            {
-                UpdateProperty(ref _exit, value);
-            }
-        }
-
-        #endregion
-
-        #region Extras Menu
-
-        private string _extras;
-        public string Extras
-        {
-            get { return _extras; }
-            set
-            {
-                UpdateProperty(ref _extras, value);
-            }
-        }
-        private string _normalize;
-        public string Normalize
-        {
-            get { return _normalize; }
-            set
-            {
-                UpdateProperty(ref _normalize, value);
-            }
-        }
-        private string _resetZoom;
-        public string ResetZoom
-        {
-            get { return _resetZoom; }
-            set
-            {
-                UpdateProperty(ref _resetZoom, value);
-            }
-        }
-        private string _registerFileExtension;
-        public string RegisterFileExtension
-        {
-            get { return _registerFileExtension; }
-            set
-            {
-                UpdateProperty(ref _registerFileExtension, value);
-            }
-        }
-        private string _unregisterFileExtension;
-        public string UnregisterFileExtension
-        {
-            get { return _unregisterFileExtension; }
-            set
-            {
-                UpdateProperty(ref _unregisterFileExtension, value);
-            }
-        }
-
-        #endregion
-
-        #region Export Menu
-
-        private string _export;
-        public string Export
-        {
-            get { return _export; }
-            set
-            {
-                UpdateProperty(ref _export, value);
-            }
-        }
-        private string _exportImage;
-        public string ExportImage
-        {
-            get { return _exportImage; }
-            set
-            {
-                UpdateProperty(ref _exportImage, value);
-            }
-        }
-        private string _useCurrentZoomOnExportedImage;
-        public string UseCurrentZoomOnExportedImage
-        {
-            get { return _useCurrentZoomOnExportedImage; }
-            set
-            {
-                UpdateProperty(ref _useCurrentZoomOnExportedImage, value);
-            }
-        }
-        private string _renderSelectionHighlightsOnExportedImage;
-        public string RenderSelectionHighlightsOnExportedImage
-        {
-            get { return _renderSelectionHighlightsOnExportedImage; }
-            set
-            {
-                UpdateProperty(ref _renderSelectionHighlightsOnExportedImage, value);
-            }
-        }
-
-        #endregion
-
-        #region Language Menu
-
-        private string _language;
-        public string Language
-        {
-            get { return _language; }
-            set
-            {
-                UpdateProperty(ref _language, value);
-            }
-        }
-
-        #endregion
-
-        #region Manage Stats Menu
-
-        private string _ManageStats;
-        public string ManageStats
-        {
-            get { return _ManageStats; }
-            set
-            {
-                UpdateProperty(ref _ManageStats, value);
-            }
-        }
-        private string _showStats;
-        public string ShowStats
-        {
-            get { return _showStats; }
-            set
-            {
-                UpdateProperty(ref _showStats, value);
-            }
-        }
-        private string _BuildingCount;
-        public string BuildingCount
-        {
-            get { return _BuildingCount; }
-            set
-            {
-                UpdateProperty(ref _BuildingCount, value);
-            }
-        }
-
-        #endregion
-
-        #region Help Menu
-
-        private string _help;
-        public string Help
-        {
-            get { return _help; }
-            set
-            {
-                UpdateProperty(ref _help, value);
-            }
-        }
-
-        private string _version;
-        public string Version
-        {
-            get { return _version; }
-            set
-            {
-                UpdateProperty(ref _version, value);
-            }
-        }
-
-        private string _versionValue;
-        public string VersionValue
-        {
-            get { return _versionValue; }
-            set
-            {
-                UpdateProperty(ref _versionValue, value);
-            }
-        }
-
-        private string _fileVersion;
-        public string FileVersion
-        {
-            get { return _fileVersion; }
-            set
-            {
-                UpdateProperty(ref _fileVersion, value);
-            }
-        }
-
-        private string _fileVersionValue;
-        public string FileVersionValue
-        {
-            get { return _fileVersionValue; }
-            set
-            {
-                UpdateProperty(ref _fileVersionValue, value);
-            }
-        }
-
-        private string _presetsVersion;
-        public string PresetsVersion
-        {
-            get { return _presetsVersion; }
-            set
-            {
-                UpdateProperty(ref _presetsVersion, value);
-            }
-        }
-
-        private string _presetsVersionValue;
-        public string PresetsVersionValue
-        {
-            get { return _presetsVersionValue; }
-            set
-            {
-                UpdateProperty(ref _presetsVersionValue, value);
-            }
-        }
-
-        private string _checkForUpdates;
-        public string CheckForUpdates
-        {
-            get { return _checkForUpdates; }
-            set
-            {
-                UpdateProperty(ref _checkForUpdates, value);
-            }
-        }
-
-        private string _enableAutomaticUpdateCheck;
-        public string EnableAutomaticUpdateCheck
-        {
-            get { return _enableAutomaticUpdateCheck; }
-            set
-            {
-                UpdateProperty(ref _enableAutomaticUpdateCheck, value);
-            }
-        }
-
-        private string _goToProjectHomepage;
-        public string GoToProjectHomepage
-        {
-            get { return _goToProjectHomepage; }
-            set
-            {
-                UpdateProperty(ref _goToProjectHomepage, value);
-            }
-        }
-
-        private string _openWelcomePage;
-        public string OpenWelcomePage
-        {
-            get { return _openWelcomePage; }
-            set
-            {
-                UpdateProperty(ref _openWelcomePage, value);
-            }
-        }
-
-        private string _aboutAnnoDesigner;
-        public string AboutAnnoDesigner
-        {
-            get { return _aboutAnnoDesigner; }
-            set
-            {
-                UpdateProperty(ref _aboutAnnoDesigner, value);
-            }
-        }
-
-        #endregion
-
-        #region Other options
-
-        private string _showGrid;
-        public string ShowGrid
-        {
-            get { return _showGrid; }
-            set
-            {
-                UpdateProperty(ref _showGrid, value);
-            }
-        }
-        private string _showLabels;
-        public string ShowLabels
-        {
-            get { return _showLabels; }
-            set
-            {
-                UpdateProperty(ref _showLabels, value);
-            }
-        }
-        private string _showIcons;
-        public string ShowIcons
-        {
-            get { return _showIcons; }
-            set
-            {
-                UpdateProperty(ref _showIcons, value);
-            }
-        }
-
-        #endregion
-
-        #region DockPanel       
-
-        private string _none;
-        public string None
-        {
-            get { return _none; }
-            set
-            {
-                UpdateProperty(ref _none, value);
-            }
-        }
-
-        private string _both;
-        public string Both
-        {
-            get { return _both; }
-            set
-            {
-                UpdateProperty(ref _both, value);
-            }
-        }
-
-        private string _search;
-        public string Search
-        {
-            get { return _search; }
-            set
-            {
-                UpdateProperty(ref _search, value);
-            }
-        }
-        private string _treeViewSearchToolTip;
-        public string TreeViewSearchToolTip
-        {
-            get { return _treeViewSearchToolTip; }
-            set
-            {
-                UpdateProperty(ref _treeViewSearchToolTip, value);
-            }
-        }
-        private string _treeViewSearchText;
-        public string TreeViewSearchText
-        {
-            get { return _treeViewSearchText; }
-            set
-            {
-                UpdateProperty(ref _treeViewSearchText, value);
-            }
-        }
-
-        #endregion
-
-        #region Status Bar
-
-        private string _statusBarControls;
-        public string StatusBarControls
-        {
-            get { return _statusBarControls; }
-            set
-            {
-                UpdateProperty(ref _statusBarControls, value);
-            }
-        }
-
-        private string _statusBarItemsOnClipboard;
-        public string StatusBarItemsOnClipboard
-        {
-            get { return _statusBarItemsOnClipboard; }
-            set
-            {
-                UpdateProperty(ref _statusBarItemsOnClipboard, value);
-            }
-        }
-
-        #endregion
-
-        private StatisticsViewModel _statisticsViewModel;
-        public StatisticsViewModel StatisticsViewModel
-        {
-            get { return _statisticsViewModel; }
-            set { _statisticsViewModel = value; }
-        }
-
-        private BuildingSettingsViewModel _buildingSettingsViewModel;
-        public BuildingSettingsViewModel BuildingSettingsViewModel
-        {
-            get { return _buildingSettingsViewModel; }
-            set { _buildingSettingsViewModel = value; }
-        }
-
-        private PresetsTreeViewModel _presetTreeViewModel;
-        public PresetsTreeViewModel PresetTreeViewModel
-        {
-            get { return _presetTreeViewModel; }
-            set { _presetTreeViewModel = value; }
-        }
-    }
-
-    public class Welcome : Notify
-    {
-        //Generated from:
-        //...
-        //public string Prop1 {get; set;}
-        //public string Prop2 {get; set;}
-        //...
-        //find expr: public (string) (.+?) {.+
-        //With the following regex (in a compatible editor that supports lowercasing of values
-        //within regex expressions):
-        //private $1 _\l$2; \r\n public $1 $2 \r\n { \r\n get { return _\l$2; } \r\n set \r\n { \r\n UpdateProperty\(ref _\l$2, value\); \r\n}\r\n}
-
-        private string _continue;
-        public string Continue
-        {
-            get { return _continue; }
-            set
-            {
-                UpdateProperty(ref _continue, value);
-            }
-        }
-
-        private string _selectALanguageWarning;
-        public string SelectALanguageWarning
-        {
-            get { return _selectALanguageWarning; }
-            set
-            {
-                UpdateProperty(ref _selectALanguageWarning, value);
-            }
-        }
-    }
-
-    public class SupportedLanguage
-    {
-        public string Name { get; set; }
-        public string FlagPath { get; set; }
     }
 }
