+++ AnnoDesigner/MainWindow.xaml
@@ -1,14 +1,21 @@
 ﻿<Window x:Class="AnnoDesigner.MainWindow"
+        x:Name="mainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
         xmlns:my="clr-namespace:AnnoDesigner"
         xmlns:coreModels="clr-namespace:AnnoDesigner.Core.Models;assembly=AnnoDesigner.Core"
-        xmlns:l="clr-namespace:AnnoDesigner.Localization"
         xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
         xmlns:converter="clr-namespace:AnnoDesigner.Core.Converter;assembly=AnnoDesigner.Core"
         xmlns:modelPresetsTree="clr-namespace:AnnoDesigner.model.PresetsTree"
         xmlns:properties="clr-namespace:AnnoDesigner.Properties"
+        xmlns:viewModel="clr-namespace:AnnoDesigner.viewmodel"
+        mc:Ignorable="d"
+        d:DataContext="{d:DesignInstance Type=viewModel:MainViewModel, IsDesignTimeCreatable=True}"
+        d:DesignHeight="480"
+        d:DesignWidth="640"
         Title="Anno Designer"
         Icon="/AnnoDesigner;component/icon.ico"
         Height="{Binding Source={x:Static properties:Settings.Default},Path=MainWindowHeight, Mode=TwoWay}"
@@ -41,11 +48,7 @@
         <HierarchicalDataTemplate DataType="{x:Type coreModels:AnnoObject}">
             <TextBlock Text="{Binding Label}" />
         </HierarchicalDataTemplate>
-        <l:MainWindow x:Key="localization"></l:MainWindow>
     </Window.Resources>
-    <Window.DataContext>
-        <Binding Source="{StaticResource localization}"></Binding>
-    </Window.DataContext>
     <xctk:BusyIndicator x:Name="busyIndicator">
         <Grid>
             <Grid.RowDefinitions>
@@ -107,7 +110,8 @@
                               Click="MenuLoadLayoutFromJsonClick" />
                     <Separator />
                     <MenuItem Header="{Binding Path=Exit}"
-                              Click="MenuItemCloseClick" />
+                              Command="{Binding CloseWindowCommand}"
+                              CommandParameter="{Binding ElementName=mainWindow}" />
                 </MenuItem>
                 <MenuItem Header="{Binding Path=Extras}"
                           TabIndex="21">
@@ -230,7 +234,7 @@
                               IsChecked="{Binding Source={x:Static properties:Settings.Default},Path=EnableAutomaticUpdateCheck, Mode=TwoWay}"></MenuItem>
                     <Separator />
                     <MenuItem Header="{Binding Path=GoToProjectHomepage}"
-                              Click="MenuItemHomepageClick" />
+                              Command="{Binding OpenProjectHomepageCommand}" />
                     <MenuItem Header="{Binding Path=OpenWelcomePage}"
                               Click="MenuItemOpenWelcomeClick"></MenuItem>
                     <MenuItem Header="{Binding Path=AboutAnnoDesigner}"
@@ -614,22 +618,27 @@
                     </StackPanel>
                 </GroupBox>
 
-                <GroupBox Header="{Binding Path=Search}"
+                <GroupBox x:Name="GroupBoxSearchPresets"
+                          Header="{Binding PresetsTreeSearchViewModel.TextSearch}"
                           HeaderStringFormat=" {0}:"
-                          Name="GroupBoxSearchPresets"
                           IsEnabled="True"
                           Height="Auto"
                           HorizontalAlignment="Stretch"
                           DockPanel.Dock="Top">
                     <Grid>
-                        <TextBox Grid.Column="0"
+                        <Grid.RowDefinitions>
+                            <RowDefinition Height="Auto" />
+                            <RowDefinition Height="Auto" />
+                        </Grid.RowDefinitions>
+                        <TextBox x:Name="TextBoxSearchPresets"
+                                 Grid.Column="0"
+                                 Grid.Row="0"
                                  Height="23"
-                                 x:Name="TextBoxSearchPresets"
                                  KeyUp="TextBoxSearchPresetsKeyUp"
                                  GotFocus="TextBoxSearchPresetsGotFocus"
                                  VerticalContentAlignment="Center"
-                                 ToolTip="{Binding TreeViewSearchToolTip}"
-                                 Text="{Binding TreeViewSearchText, UpdateSourceTrigger=PropertyChanged}">
+                                 ToolTip="{Binding PresetsTreeSearchViewModel.TextSearchToolTip}"
+                                 Text="{Binding PresetsTreeSearchViewModel.SearchText, UpdateSourceTrigger=PropertyChanged}">
                         </TextBox>
                         <!-- Maybe add a clear text box button eventually. (Needs styling etc)-->
                         <!--<Button Grid.Column="1"
@@ -641,6 +650,10 @@
                             Background="Transparent"
                             BorderThickness="0"
                             />-->
+                        <xctk:CheckComboBox Grid.Column="0"
+                                            Grid.Row="1"
+                                            IsEditable="False"
+                                            Visibility="Collapsed" />
                     </Grid>
                 </GroupBox>
                 <GroupBox x:Name="GroupBoxPresets"
@@ -649,21 +662,21 @@
                           Width="{Binding ElementName=BuildingsPanel, Path=ActualWidth, Mode=OneWay}"
                           DockPanel.Dock="Top">
                     <TreeView x:Name="PresetsTree"
-                              ItemsSource="{Binding PresetTreeViewModel.FilteredItems}"
+                              ItemsSource="{Binding PresetsTreeViewModel.FilteredItems}"
                               VirtualizingStackPanel.IsVirtualizing="True"
                               VirtualizingStackPanel.VirtualizationMode="Recycling">
                         <TreeView.InputBindings>
-                            <KeyBinding Command="{Binding PresetTreeViewModel.ReturnKeyPressedCommand}"
+                            <KeyBinding Command="{Binding PresetsTreeViewModel.ReturnKeyPressedCommand}"
                                         Key="Return"
                                         CommandParameter="{Binding SelectedItem, ElementName=PresetsTree}" />
                         </TreeView.InputBindings>
                         <i:Interaction.Triggers>
                             <i:EventTrigger EventName="MouseDoubleClick">
-                                <i:InvokeCommandAction Command="{Binding PresetTreeViewModel.DoubleClickCommand}"
+                                <i:InvokeCommandAction Command="{Binding PresetsTreeViewModel.DoubleClickCommand}"
                                                        CommandParameter="{Binding SelectedItem, ElementName=PresetsTree}" />
                             </i:EventTrigger>
                             <!--<i:EventTrigger EventName="SelectedItemChanged">
-                                <i:InvokeCommandAction Command="{Binding PresetTreeViewModel.SelectedItemChangedCommand}"
+                                <i:InvokeCommandAction Command="{Binding PresetsTreeViewModel.SelectedItemChangedCommand}"
                                                        CommandParameter="{Binding SelectedItem, ElementName=PresetsTree}" />
                             </i:EventTrigger>-->
                         </i:Interaction.Triggers>
