+++ AnnoDesigner/Welcome.xaml.cs
@@ -1,4 +1,5 @@
-﻿using AnnoDesigner.Localization;
+﻿using AnnoDesigner.Core.Models;
+using AnnoDesigner.model;
 using AnnoDesigner.Properties;
 using System;
 using System.Collections.Generic;
@@ -20,70 +21,11 @@ namespace AnnoDesigner
     /// <summary>
     /// Interaction logic for Welcome.xaml
     /// </summary>
-    public partial class Welcome : Window
+    public partial class Welcome : Window, ICloseable
     {
         public Welcome()
         {
             InitializeComponent();
-            var languages = new List<SupportedLanguage>()
-            {
-                new SupportedLanguage()
-                {
-                    Name = "English",
-                    FlagPath = "Flags/United Kingdom.png"
-                },
-                new SupportedLanguage()
-                {
-                    Name = "Deutsch",
-                    FlagPath = "Flags/Germany.png"
-                },
-                new SupportedLanguage()
-                {
-                    Name = "Français",
-                    FlagPath = "Flags/France.png"
-                },
-                new SupportedLanguage()
-                {
-                    Name = "Polski",
-                    FlagPath = "Flags/Poland.png"
-                },
-                new SupportedLanguage()
-                {
-                    Name = "Русский",
-                    FlagPath = "Flags/Russia.png"
-                }
-            };
-            LanguageSelection.ItemsSource = languages;
-        }
-
-        private void LanguageSelection_MouseDoubleClick(object sender, MouseButtonEventArgs e)
-        {
-            var lb = (ListBox)sender;
-            if (lb.SelectedItem != null)
-            {
-                LoadSelectedLanguage();
-            }
-        }
-
-        private void ContinueButton_Click(object sender, RoutedEventArgs e)
-        {
-            if (LanguageSelection.SelectedItem != null)
-            {
-                LoadSelectedLanguage();
-            }
-            else
-            {
-                //Show a message;
-                MessageBox.Show(this, "Please select a langauge before continuing");
-            }
-        }
-
-        private void LoadSelectedLanguage()
-        {
-            MainWindow.SelectedLanguage = ((SupportedLanguage)LanguageSelection.SelectedItem).Name;
-            Settings.Default.SelectedLanguage = MainWindow.SelectedLanguage;
-            Settings.Default.Save();
-            this.Close();
         }
     }
 }
