+++ AnnoDesigner/model/StatisticsBuilding.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using AnnoDesigner.Core.Models;
 
 namespace AnnoDesigner.model
 {
