+++ AnnoDesigner/viewmodel/BuildingSettingsViewModel.cs
@@ -1,6 +1,5 @@
 using AnnoDesigner.Core.Models;
 using AnnoDesigner.Core.Presets.Helper;
-using AnnoDesigner.model;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
